server/client-server/requirements.txt
requests
Flask>=1.1.1
flask-restplus>=0.13.0
opencv-python-headless
Pillow>=7.1
alphashape==1.1.0
waitress==1.4.4
fastapi>=0.54.1
uvicorn==0.11.5
werkzeug==0.16.1
shapely==1.7.1
paste==3.5.0
click>=7.0
matplotlib
python-magic
aiofiles
loguru
tritonclient[all]
